+++ engine_tiflash/src/ps_engine/mod.rs
@@ -1,0 +1,4 @@
+// Copyright 2022 TiKV Project Authors. Licensed under Apache-2.0.
+
+mod ps_db_vector;
+pub use ps_db_vector::*;
